+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Get-MyRMInventoryInfo` shows whether the inventory file exists or not
 - Check and warn if duplicates exist at inventory file access
 - `Test-MyRMConnection` Cmdlet to test a connection defined in the inventory
+- `RunInCurrentScope` parameter (alias `X`) in `Invoke-MyRMConnection` to start the connection process in the current console
 
 ### Changed